+++ src/events/event_handler.ts
@@ -13,11 +13,14 @@ import { diff_match_patch } from 'diff-match-patch';
 import { VSCODE } from "../constants";
 import { isRepoSynced } from './utils';
 import { removeFile } from '../utils/file_utils';
+import { CODESYNC_STATES, CodeSyncState } from '../utils/state_utils';
+import gitCommitInfo from 'git-commit-info';
 
 
 export class eventHandler {
 	repoPath: string;
 	branch: string;
+	commitHash: string|null;
 	viaDaemon: boolean;
 	repoIsNotSynced: boolean;
 	pathUtils: any;
@@ -46,6 +49,16 @@ export class eventHandler {
 		this.originalsRepoBranchPath = this.pathUtils.getOriginalsRepoBranchPath();
 		this.syncIgnoreItems = getSyncIgnoreItems(this.repoPath);
 		this.defaultIgnorePatterns = getDefaultIgnorePatterns();
+		if (viaDaemon) {
+			const commitInfo = gitCommitInfo({cwd: this.repoPath});
+			this.commitHash = commitInfo.hash || null;
+		} else {
+			this.commitHash = CodeSyncState.get(CODESYNC_STATES.GIT_COMMIT_HASH) || null;
+			if (!this.commitHash) {
+				const commitInfo = gitCommitInfo({cwd: this.repoPath});
+				this.commitHash = commitInfo.hash || null;	
+			}
+		}
 	}
 
 	addDiff = (relPath: string, diffs="") => {
@@ -59,6 +72,7 @@ export class eventHandler {
 		newDiff.source = VSCODE;
 		newDiff.repo_path = this.repoPath;
 		newDiff.branch = this.branch;
+		newDiff.commit_hash = this.commitHash;
 		newDiff.file_relative_path = relPath;
 		newDiff.diff = diffs;
 		newDiff.is_new_file = this.isNewFile;