+++ CHANGELOG.md
@@ -4,8 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.32.0] - 2023-11-04
-- Minor tweaks in Login Success and Failed HTML files
+## [3.35.0] - 2023-12-23
+- Reactivate Account functionality added if user has deactivated the account from webapp
+- Improved logic for s3Uploader
+
+## [3.34.2] - 2023-12-14
+- Getting commit hashes for populate-buffer for each repo
+
+## [3.34.1] - 2023-12-14
+- Fix for Window path
+
+## [3.34.0] - 2023-12-14
+- Sending commit_hashes along with diffs to the server
+
+## [3.33.0] - 2023-11-04
+- Mionr tweaks in Login Success and Failed HTML files
 
 ## [3.31.0] - 2023-08-16
 - s3Uploader introduced to manage repo, branch and file upload to s3
@@ -13,7 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Error Handling: If some chunk is failed to upload successfully somehow, 
     we save that chunk in a separate file to retry later by reducing chunk size
 - Shows a spinner loader in status bar while branch is being uploaded to the server
-
 
 ## [3.30.1] - 2023-08-08
 - Fixed passing AWS Config to CloudWatchLogsClient